JAMES-2754 extract external James configuration
